pyproject.toml
[tool.poetry]
name = "kedro-vertexai"
version = "0.7.0"
description = "Kedro plugin with GCP Vertex AI support"
readme = "README.md"
authors = ['Marcin Zabłocki <marcin.zablocki@getindata.com>', 'Mateusz Pytel <mateusz.pytel@getindata.com>', 'Mariusz Strzelecki <mariusz.strzelecki@getindata.com>']
maintainers = ['GetInData MLOPS <mlops@getindata.com>']
homepage = "https://github.com/getindata/kedro-vertexai"
repository = "https://github.com/getindata/kedro-vertexai"
documentation = "https://kedro-vertexai.readthedocs.io/"
keywords = ['kedro-plugin', 'kedro', 'mlops', 'vertexai', 'googlecloudplatform', 'machinelearning']
license = "Apache-2.0"
classifiers = [
    "Development Status :: 4 - Beta", # license and python versions added automatically
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
branch = true
relative_files = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError"
]

[tool.isort]
known_third_party = ["click","google","kedro","kfp","kubernetes","tabulate", "pydantic","semver","setuptools"]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
kedro = ">=0.18.1, <0.19.0"
click = ">=8.0.4"
kfp = ">=1.8.12,<2.0"
tabulate = ">=0.8.7"
semver = "~=2.10"
pydantic = "~=1.9.0"
google-auth = "<3"
google-cloud-scheduler = ">=2.3.2"
google-cloud-iam = "<3"
gcsfs = "<=2022.1,>=2021.4"
fsspec = "<=2022.1,>=2021.4"
google-cloud-storage = "<3.0.0"
grpcio = "~=1.44.0"
grpcio-status = "~=1.44.0"
protobuf = "<=3.20.0"
kedro-mlflow = { version = "~=0.11.1", optional = true }
cachetools = ">=4.1,<5.0"

[tool.poetry.extras]
mlflow = ["kedro-mlflow"]

[tool.poetry.dev-dependencies]
pytest = "<7"
pytest-cov = ">=2.8.0, <4.0.0"
tox = ">=3.25.1"
pre-commit = "2.20.0"
pytest-subtests = ">=0.7.0, <1.0.0"
responses = ">=0.13.4"

[tool.poetry.plugins] # Optional super table
[tool.poetry.plugins."kedro.project_commands"]
"vertexai" = "kedro_vertexai.cli:commands"

[tool.poetry.plugins."kedro.hooks"]
"vertexai_mlflow_tags_hook" = "kedro_vertexai.hooks:mlflow_tags_hook"
"vertexai_cfg_hook" = "kedro_vertexai.hooks:env_templated_config_loader_hook"

[tool.poetry.plugins."mlflow.request_header_provider"]
"unused" = "kedro_vertexai.auth.mlflow_request_header_provider:DynamicMLFlowRequestHeaderProvider"
